Moved the League of Legends information to its own child component
