--- src/app/video-games/video-games.component.scss
@@ -1,103 +1,0 @@
-.rank-container {
-  @apply flex flex-col items-center justify-center py-4 px-10 sm:px-20 md:px-10;
-  border: 1px solid rgba(var(--lavender), 0.2);
-
-  // Avoid double borders
-  &:not(:last-of-type) {
-    border-bottom: 0;
-  }
-
-  @media (min-width: theme('screens.md')) {
-    &:not(:last-of-type) {
-      border-right: 0;
-      border-bottom: 1px solid rgba(var(--lavender), 0.2)
-    }
-  }
-}
-
-.league-of-legends-box {
-  background: #1c4a5f;
-  background: linear-gradient(135deg, #1c4a5f 35%, #55551b 100%);
-  border: 1px solid #121720;
-  @apply rounded-lg p-4 pb-8 mx-3;
-}
-
-.profile-icon {
-  border: 2px solid rgb(34 110 125);
-}
-
-// Adapted from the travel component
-.tooltip {
-  visibility: hidden;
-  opacity: 0;
-  position: absolute;
-  z-index: 5;
-  top: -2.75rem;
-  left: 50%;
-  translate: -50%;
-  text-align: center;
-  background-color: rgb(var(--charcoal));
-  color: rgb(var(--lavender));
-  border-radius: 4px;
-  @apply py-1.5 px-3;
-  text-wrap: nowrap;
-  white-space: nowrap;
-  transition: opacity 0.2s linear;
-
-  &::after { // Reference: https://www.w3schools.com/css/css_tooltip.asp
-    content: " ";
-    position: absolute;
-    top: 100%;
-    left: 50%;
-    margin-left: -5px;
-    border-width: 5px;
-    border-style: solid;
-    border-color: rgb(var(--charcoal)) transparent transparent transparent;
-  }
-}
-
-.champion-icons-container {
-  @apply grid justify-center items-center gap-2 max-w-2xl px-4 mx-auto;
-  grid-template-areas:
-    'first'
-    'second'
-    'third'
-    'fourth'
-    'fifth';
-
-  @media (min-width: theme('screens.2xs')) {
-    grid-template-areas:
-    'first first second second third third'
-    '. fourth fourth fifth fifth .';
-  }
-
-  @media (min-width: theme('screens.sm')) {
-    grid-template-areas:
-      'first second third fourth fifth';
-  }
-
-  .champion-icon:nth-child(1) {
-    grid-area: first;
-  }
-
-  .champion-icon:nth-child(2) {
-    grid-area: second;
-  }
-
-  .champion-icon:nth-child(3) {
-    grid-area: third;
-  }
-
-  .champion-icon:nth-child(4) {
-    grid-area: fourth;
-  }
-
-  .champion-icon:nth-child(5) {
-    grid-area: fifth;
-  }
-}
-
-.champion-icon:hover .tooltip {
-  visibility: visible;
-  opacity: 1;
-}
